fix(Body) from Int to Double

+++ src/main/scala/it/unibo/pps/model/body/Body.scala
@@ -1,5 +1,5 @@
 package it.unibo.pps.model.body
 
-case class Body(position: Tuple2[Int, Int], velocity: Tuple2[Int, Int])
+case class Body(position: Tuple2[Double, Double], velocity: Tuple2[Double, Double])
 
 object Body:
